Added doesFileExist checks before read or remove logged files

+++ dhis2-android-app/src/main/java/org/dhis2/mobile/utils/TextFileUtils.java
@@ -197,7 +197,9 @@ public final class TextFileUtils {
     public static void removeFile(Context context, Directory dir, FileNames fileName) {
         String path = getDirectoryPath(context, dir);
         File file = new File(path, fileName.toString());
-        removeFile(file);
+        if(file.exists()) {
+            removeFile(file);
+        }
     }
 
     public static void removeFile(File file) {
